run forrest run
